requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
aiosqlite==0.17.0
asgiref==3.8.1
asyncpraw==7.8.0
asyncprawcore==2.4.0
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
Django==5.1.4
django-crispy-forms==2.3
django-environ==0.11.2
django-haystack==3.3.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
mysqlclient==2.2.6
packaging==24.2
praw==7.8.1
prawcore==2.4.0
propcache==0.2.1
pysolr==3.10.0
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.3
websocket-client==1.8.0
yarl==1.18.3
